Builds on odroid

+++ src/ext/OpenMVS/Interface.h
@@ -216,7 +216,7 @@ namespace _INTERFACE_NAMESPACE {
 
 		// Serialization support for basic types
 		ARCHIVE_DEFINE_TYPE(uint32_t)
-		ARCHIVE_DEFINE_TYPE(size_t)
+		ARCHIVE_DEFINE_TYPE(uint64_t)
 		ARCHIVE_DEFINE_TYPE(float)
 		ARCHIVE_DEFINE_TYPE(double)
